feat: condensed ui for comment awarding
